Removes unecessary setup code from FirebaseSessions and SessionInitiator.
