+++ firebase-sessions/src/main/kotlin/com/google/firebase/sessions/FirebaseSessions.kt
@@ -16,49 +16,29 @@
 
 package com.google.firebase.sessions
 
+import android.app.Activity
 import android.app.Application
+import android.app.Application.ActivityLifecycleCallbacks
+import android.os.Bundle
 import android.util.Log
 import com.google.firebase.Firebase
 import com.google.firebase.FirebaseApp
 import com.google.firebase.app
 import com.google.firebase.sessions.api.FirebaseSessionsDependencies
 import com.google.firebase.sessions.api.SessionSubscriber
-import com.google.firebase.sessions.settings.SessionsSettings
-import kotlinx.coroutines.CoroutineDispatcher
 
 /** The [FirebaseSessions] API provides methods to register a [SessionSubscriber]. */
-class FirebaseSessions
-internal constructor(
-  private val firebaseApp: FirebaseApp,
-  internal val backgroundDispatcher: CoroutineDispatcher,
-  private val sessionFirelogPublisher: SessionFirelogPublisher,
-  sessionGenerator: SessionGenerator,
-  private val sessionSettings: SessionsSettings,
-) {
-
-  // TODO: This needs to be moved into the service to be consistent across multiple processes.
-  private val collectEvents: Boolean
+class FirebaseSessions internal constructor(private val firebaseApp: FirebaseApp) {
 
   init {
-    collectEvents = shouldCollectEvents()
-
-    val sessionInitiator =
-      SessionInitiator(
-        timeProvider = WallClock,
-        backgroundDispatcher,
-        ::initiateSessionStart,
-        sessionSettings,
-        sessionGenerator,
-      )
-
     val appContext = firebaseApp.applicationContext.applicationContext
     if (appContext is Application) {
       SessionLifecycleClient.bindToService(appContext)
-      appContext.registerActivityLifecycleCallbacks(sessionInitiator.activityLifecycleCallbacks)
+      appContext.registerActivityLifecycleCallbacks(activityLifecycleCallbacks)
 
       firebaseApp.addLifecycleEventListener { _, _ ->
         Log.w(TAG, "FirebaseApp instance deleted. Sessions library will not collect session data.")
-        appContext.unregisterActivityLifecycleCallbacks(sessionInitiator.activityLifecycleCallbacks)
+        appContext.unregisterActivityLifecycleCallbacks(activityLifecycleCallbacks)
       }
     } else {
       Log.e(
@@ -79,50 +59,7 @@ internal constructor(
     )
   }
 
-  private suspend fun initiateSessionStart(sessionDetails: SessionDetails) {
-    val subscribers = FirebaseSessionsDependencies.getRegisteredSubscribers()
-
-    if (subscribers.isEmpty()) {
-      Log.d(
-        TAG,
-        "Sessions SDK did not have any dependent SDKs register as dependencies. Events will not be sent."
-      )
-      return
-    }
-
-    subscribers.values.forEach { subscriber ->
-      // Notify subscribers, regardless of sampling and data collection state.
-      subscriber.onSessionChanged(SessionSubscriber.SessionDetails(sessionDetails.sessionId))
-    }
-
-    if (subscribers.values.none { it.isDataCollectionEnabled }) {
-      Log.d(TAG, "Data Collection is disabled for all subscribers. Skipping this Session Event")
-      return
-    }
-
-    Log.d(TAG, "Data Collection is enabled for at least one Subscriber")
-
-    // This will cause remote settings to be fetched if the cache is expired.
-    sessionSettings.updateSettings()
-
-    if (!sessionSettings.sessionsEnabled) {
-      Log.d(TAG, "Sessions SDK disabled. Events will not be sent.")
-      return
-    }
-
-    if (!collectEvents) {
-      Log.d(TAG, "Sessions SDK has dropped this session due to sampling.")
-      return
-    }
-  }
-
   /** Calculate whether we should sample events using [sessionSettings] data. */
-  private fun shouldCollectEvents(): Boolean {
-    // Sampling rate of 1 means the SDK will send every event.
-    val randomValue = Math.random()
-    return randomValue <= sessionSettings.samplingRate
-  }
-
   companion object {
     private const val TAG = "FirebaseSessions"
 
@@ -140,6 +77,27 @@ internal constructor(
         app.get(FirebaseSessions::class.java)
       } else {
         throw IllegalArgumentException("Firebase Sessions only supports the Firebase default app.")
+      }
+
+    /**
+     * Lifecycle callbacks that will inform the [SessionLifecycleClient] whenever an [Activity] in
+     * this application process goes foreground or background.
+     */
+    private val activityLifecycleCallbacks =
+      object : ActivityLifecycleCallbacks {
+        override fun onActivityResumed(activity: Activity) = SessionLifecycleClient.foregrounded()
+
+        override fun onActivityPaused(activity: Activity) = SessionLifecycleClient.backgrounded()
+
+        override fun onActivityCreated(activity: Activity, savedInstanceState: Bundle?) = Unit
+
+        override fun onActivityStarted(activity: Activity) = Unit
+
+        override fun onActivityStopped(activity: Activity) = Unit
+
+        override fun onActivityDestroyed(activity: Activity) = Unit
+
+        override fun onActivitySaveInstanceState(activity: Activity, outState: Bundle) = Unit
       }
   }
 }
