--- firebase-sessions/src/main/kotlin/com/google/firebase/sessions/SessionInitiateListener.kt
@@ -1,23 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.firebase.sessions
-
-/** Interface for listening to the initiation of a new session. */
-internal fun interface SessionInitiateListener {
-  /** To be called whenever a new session is initiated. */
-  suspend fun onInitiateSession(sessionDetails: SessionDetails)
-}
